python/requirements.txt
certifi==2020.6.20
cffi==1.15.1
coverage==4.5.4
django-db-connection-pool==1.2.2
enum34==1.1.10
gevent==22.10.2
gunicorn==19.10.0
http==0.2
idna==2.10
ipaddress==1.0.23
mock==3.0.5
mysql-connector-python==8.0.32
mysqlclient==1.4.6
pip-chill==1.0.1
pyasn1==0.4.8
pyjwt==1.7.1
pyparsing==2.4.7
python==2.7.18
python-dotenv==0.18.0
south==1.0.2
wsgiref==0.1.2
